Only let users who are fraud review pending visit the please call controller (#8812)
The please call controller is used to instruct users who are pending fraud review to contact Login.gov to finish verification.

This controller should only be accessible to users who are fraud review. This commit adds a before action to redirect users in the following scenarios:

1. The user is not fraud review pending
2. The user has been fraud rejected (these users get a different message rendered by a different controller)

This commit also fixes a latent 500 error on this controller. A user who signs in but is not fraud pending would encounter a `NoMethodError` at this line:

```ruby
pending_at = current_user.fraud_review_pending_profile.fraud_review_pending_at
```

[skip changelog]