LG-10289: Fix cropping error message
[Ticket](https://cm-jira.usa.gov/browse/LG-10289)

Before this change, when the camera cropping failed in the SDK, the user
would get a `Camera failed to start, please try again.` error message.

This is confusing because to get to the cropping failure, the user's
camera would have had to start. In the case of a cropping error, we want
to give a general error message instead.