Reduce active / pending profile down to a single current profile
Currently a user can an active profile while also have a pending profile. This covers an edge case where a user who has already undergone proofing is making a subsequent attmept. Maintaining the active profile while their subsequent profile is pending means they maintain an active verified account while completing out-of-band steps for their second profile.

This edge case should be relatively rare, but does introduce a lot of complexity to the application.

This commit removes the active and pending profile model in favor of using a `current_profile` that is the most recent profile. As a result, a user who undergoes proofing again with an active profile will have their active profile overwritten once their new profile is written.

This reduces the active profile vs current profile question down to a question of the state of the user's latest profile.

[skip changelog]