LG-8763 Log that a profile is pending on password reset (#8872)
* LG-8763 Log that a profile is pending on password reset

When a user resets their password it deactivates their active profile. We log that we do that.

Additionally, we invalidate any pending profiles which will encounter a encryption error when the user logs in. This commit adds logging to log that a pending profile is present and the reasons that it is pending so we can measure the impact of this behavior at password reset

[skip changelog]