LG-10439 Add `had_barcode_read_failure` to `idv_session`
We are working to retire `flow_session` as part of retiring FSM and restoring the back button functionality.

This commit starts to move one of the properties, `had_barcode_read_failure`, into `idv_session`. Another commit will be necessary to start reading from the `idv_session` and then a final commit to stop writing to `flow_session`.

[skip changelog]