Implement current_profile relation on user
This commit adds a more conservative approach to #8866.

That PR investigated what it would take to fully re-implment the pending/active/password_reset profile tooling to ensure those always returned the same profile. There are some pretty large changes that come along with that.

This commit simplifies things by keeping the interface that fetches those profiles on the user. It changes the implmentation so they only return a profile if it is the most recent profile.

[skip changelog]