LG-10439 Read bardcode read error values form the session
We are working on removing the flags that indicate a bardcode read error from the `flow_session` and adding them to `idv_session` instead.

The work to write the values to `idv_session` was handled in #8881 and #8889. That has been deployed so it is not safe to start reading the values from `idv_session`. A future commit will stop writing them to `flow_session`.

[skip changelog]