LG-10342 Introduce ciphertext group for managing single and multi-region KMS ciphertexts
We are working on migrating to a new KMS key that has multi-region capabilities. Migrating to this new key will require storing both single region and multi-region ciphertexts in the database. Moreover, the inner layer of these ciphertexts is expected to match so we cannot have 2 encryptors, 1 with the new key and 1 with the old.

This commit introduces a `RegionalCiphertextPair` struct. It is intended to be returned from the `PiiEncryptor` and eventually the `PasswordVerifier`.

Consumers of the `PiiEncryptor#encrypt` methods can break up the `RegionalCiphertextPair` and store the single region and multi-region ciphertexts independently. On decryption the ciphertext pair can be reconstructed and used as the argument to `PiiEncryptor#decrypt`.

[skip changelog]