LG-10342 Add the ability to specify a key_id for the KMS Client
We are migrating to a new KMS instance for encryption that supports multi-region keys. We will need to encrypt and decrypt ciphertexts with both the current single region key and the new multi-region key. To support that the KMS client will need to be able to be configured for one of the given keys.

This commit adds the ability to specify a key ID to use for the KMS client. To preserve the original behavior as the new key is being phased in the KMS client uses the old single region key if no key is specified.

[skip changelog]