LG-10342 Encrypt PII with the multi-region encryptor (#8958)
We are working on migrating to a multi-region capable KMS key. As part of the migration we will be temporarily writing ciphertexts encrypted with the current KMS key and the new multi-region key.

This commit makes a change to the PII encryptor such that it encrypts with the multi-region key when the `aws_kms_multi_region_write_enabled` flag is set. This will cause a multi-region ciphertext to be returned and written to the database.

[skip changelog]