LG-10342 Encrypt password digests with multi-region key
We are working on migrating to a KMS key that supports encryption and decryption across regions. To enable this move we are changing the application to encrypt PII and passwords with both a multi-region and single-region key during the migraiton.

This commit changes the app to encrypt passwords with the single-region key and the multi-region key when the `aws_kms_multi_region_write_enabled` flag is set.

[skip changelog]