Do a little dance to preserve querystring
We kind of clobber the original URL the user was going to if they are redirected as part of the sign in process (see ApplicationController::after_sign_in_path_for). So here we first check if the user was _going_ to see the "didn't get a letter" variant, store that in session, then re-redirect them to the querystring'd version of the URL after they are signed in.