Test that after max retries, see expected heading
Originally we were going to test that the tips copy wasn't shown. But
now that we're deleting the copy from the repo entirely, expecting the
tips not to be shown, in a way, would be keeping the tips in the repo.

So instead, we are are adding a test to make sure we see what we expect
on the mobile side.

Implementation note: We have to freeze time; otherwise the `timeout`
will cause the test to flicker. (Example: The test will look for
"try again in 6 hours" but the page will say "try again in 5 hours and
59 minutes".)