Modify MarketingSite.valid_help_center_article? to check if a article URI is valid
We had an issue where someone was passing invalid `article_anchor` params to the `HelpCenterController`. These params were causing the `MarketingSite` helpers to raise `URI::InvalidURIError` and render a 500.

This commit provides some extra validation to the `valid_help_center_article?` to prevent this issue. Specifically it attempts to build the article URL and rescues if it raises `URI::InvalidURIError`.

This does mean that article URLs get constructed twice which is a little annoying, but it prevents these 500s from occuring while maintaining the behavior where `help_center_article_url` always returns something or raises a breaking exception and `valid_help_center_article?` is used to guard calls to it.

[skip changelog]