Add allow_browser_log back into test
Initially I found I needed this when I had the test running in the
desktop context. But when I ran the test in the mobile context, I didn't
seem to need it.

I just saw the test flicker again with a console log error, so I am
adding it back in to guard against future flickering.