Add second template for Welcome controller for A/B test
We got unexpected results for our A/B test for the GettingStarted page that replaces both the Welcome and
Agreement steps. Add a third option that just replaces the Welcome template with one that has the new content
for the GettingStarted page (except the Agreement content) and then sends the user to the Agreement step as
usual.

changelog: User-facing Improvements, Identity Verification Getting Started Experience, Additional Welcome template for three-way A/B test

Co-authored-by: Jonathan Hooper <jonathan.hooper@gsa.gov>