LG-10518: Delete capture troubleshooting component and update tests (#9017)
changelog: Internal, Doc Auth, Delete CaptureTroubleshooting Component and update tests

We don't want to show troubleshooting tips this way any longer. Instead, we're looking to use the help center for showing this sort of advice.

In #8842, in order to avoid showing the tips, we increased the amount of tries before showing the tips to well past the maximum amount of total tries. Upon further discussion, we decided it would be better to remove the CaptureTroubleshooting component entirely, as we have no plans to reuse it in the future.

Once CaptureTroubleshooting was removed, a few more things were now unused and could also be removed - the config setting the amount of tries before tips were shown, the copy used in the tips, the CaptureAdvice component, and any place calling these things.

Includes removing imports that are now unused within the spec file and unused svg files

Implementation notes:

- we chose to [use fragment instead of div](https://medium.com/fasal-engineering/what-are-react-fragments-or-the-react-empty-tags-190253582905) when replacing the `CaptureTroubleshooting` wrapper in order to avoid unnecessary HTML elements.