Remove old fraud review profile state backfill rake tasks
These rake tasks were added as part of the profile state work. They were used to backfill new columns that were added to track the state of a profile with regards to fraud review.

The backfills have been successfully completed so we do not need these rake tasks any longer.

[skip changelog]