LG-10344 Add a rake task for detecting mismatch multi-region and single-region KMS records (#9058)
We are in the midst of a migration between KMS keys. We are changing from a single-region only key to a key that supports multi-region use.

As part of this migration we are encrypting data twice: once with each of the keys. Under no circumstance should the data that is encrypted by KMS and stored be different for a given record.

This commit introduces a rake task that confirms that the same data is being encrypted and written. It does so by grabbing a sample of records, removing the KMS encryption, and comparing the contents.

[skip changelog]