Retire `log_password_verification_failure` (#9084)
This method was added to log exceptions that were raised when trying to validate passwords. This was because an invalid password attempt would raise an encryption error; there was concern that this would cause other exceptions to be missed. This was only relevant when we were encrypting passwords with UAKs.

Now that we are not longer using UAKs this method just writes invalid password attempts to production.log. This duplicates what is written to events.log for events where passwords are entered.

I could not find anywhere that we are looking at this log line. This commit removes it.

[skip changelog]