Configure dependencies to allow cloudwatch reports to run in production
I tried generating the IdentityVerificationReport in production and learned that we do not have the AWS Cloudwatch SDK in our production group.

I addressed the above by adding it and did a quick sweep of dependencies that were in the Gemfile to add things that were missing there. I also removed some unnecessary `require`s where they were being used for things that are already loaded with all the other gems.

[skip changelog]