LG-10823 Report Faraday timeout errors to NewRelic (#9073)
Currently we do not report MVA timeouts or DLDV connection issues to NewRelic. We do this because this service often has issues and creates a lot of noise.

The issues we observe are generally the result of individual MVA timeouts and not issues connecting to the DLDV service. Moreover, during DLDV outages not having DLDV issues reported limits our visibility.

This commit removes the constraint that prevents DLDV connection issues from being reported so we have visibility. Since the DLDV service itself is generally reliable this should not result in too much noise.

[skip changelog]