LG-10836 Configure IdentityVerificationReport job to run every 24 hours (#9089)
A previous commit added a job to create an identity verification report based on cloudwatch data. The job creates the report and uploads it to S3.

Since this is the first report to leverage cloudwatch data in this way we planned on manually testing the report before setting it up to run on a recurring basis. This commit does the rest of the work to make the report run regularly after we confirm it actually works.

[skip changelog]