LG-10343 Add a class for migrating encrypted PII to the multi-region KMS key (#9114)
We are working on migrating between KMS keys. The new key will support multi-region encryption. As part of this migration we encrypting records with both keys.

Previous commits added code for encrypting with both keys and writing ciphertexts for both on `INSERT` and `UPDATE`. This commit adds tooling for going back and updating records that are not being changed by normal user behavior.

changelog: Internal, Multi-Region KMS, A multi-region KMS pii migrator was added for future use to encrypt and save encrypted pii and encrypted recovery pii columns with the new multi-region KMS key after decrypting the KMS layer with the single region KMS key.