LG-10891 Add a job for migrating user password digests to multi-region KMS
We are in the midst of a migration from a single-region KMS key to a multi-region capable KMS key. That migration involves the following steps:

1. Add new columns for multi-region ciphertexts
2. Start writing to the new columns with multi-region ciphertexts
3. Start opportunistically reading from the multi-region columns
4. Backfill the multi-region columns for old users
5. Stop using the old single-region columns

This commit is part of step number 4.

This commit adds a `UserMigrator` class which takes a user and migrates the single-region encrypted `encrypted_password_digest` and `encrypted_recovery_code_digest` to `encrypted_password_digest_multi_region` and `encrypted_recovery_code_digest_multi_region` respectivally. This commit also includes a job to find users that require migration and invoke the `UserMigrator` against them.

The background job is not currently on the background job schedule. This will allow us to do some testing with it in lower environments before we start using it in production.

changelog: Internal, Multi-region KMS migration, A UserMigrator and UserMigration job was added for finding users that have an encrypted password digest or encrypted recovery code digest where the outer layer is encrypted with a single region KMS key and decrypting the outer layer using the single-region KMS key and then encrypting the results with a multi-region capable KMS key before saving the result of the multi-region capable KMS encryption in the columns that were added for multi-region password and recovery code digests in a previous commit.