LG-10343 Enable the profile backfill job (#9162)
We have added code to start writing to profiles with a multi-region KMS key. This is in addition to the single-region key was were using previously. Records that were created before we introduced the multi-region key do not have a value encrypted with the multi-region KMS key.

https://github.com/18F/identity-idp/commit/5506d9e4dfe67283d8bd36f25df333ddd3aa9f31 added a background job for backfilling the empty multi-region ciphertexts. This commit enables it after testing and timing it in lower environments and in production.

changelog: Internal, Multi-region KMS migration, The MultiRegionKmsMigration::ProfileMigrationJob job which finds profiles with PII encrypted with the single-region KMS key and not the multi-region KMS key and decrypts the single-region KMS layer using the single region key and encrypts the result with the multi-region KMS key and saves the resulting ciphertext as the multi-region ciphertext was enabled.