LG-10347 Make the key ID the session encryptor uses configurable
We are currently migrating from a single-region KMS key to a multi-region capable KMS key.

This commit modifies the SessionEncryptor and BackgroundArgsEncryptor to have to use a configurable key for their KMS client. These encryptors are used in 3 contexts:

1. Encryption of sessions (SessionEncryptor)
2. Encryption of GPO confirmation entries (SessionEncryptor)
3. Encryption of arguments to background jobs (BackgroundArgsEncryptor)

The KMS client's for these encryptors will now use the configured key ID for encryptions. For decryption the client allows KMS to select the key. This means that decryption will not be affected by this change as long as KMS still has access to the keys referenced by the KeyID used for encryption.

Since all of the encryption operations done with these encryptors produce ephemeral ciphertexts there is not need to worry about holding onto old keys after this has been deployed with the multi-region key configured for a while.

changelog: Internal, Multi-region KMS migration, The SessionEncryptor and BackgroundArgsEncryptor were change to have a configurable KMS key ID that is used for encryption in order to facilitate a migration to a multi-region key ID.