LG-10779 Add a specific test case for a profile that cannot pass fraud review
The review profile rake task is used to pass or reject users after fraud review. We have a test case for users who running that task against a user who is not pending fraud review because they have undergone a more recent verification. Previously that test case actually tested a case where a user had an invalid profile state that was pending GPO verification. This should not be possible and raises an exception. It also did not match the spec description.

This commit cleans up the spec to cover the case described in the spec description. In this case the user has reset their password and deactivated the fraud pending profile. Then they have created a new profile that is GPO pending.

A new spec is also introduced to cover the code paths covered by the old test: a case where the profile is in an unexpected state and an exception is raised.

[skip changelog]