LG-10891 Add MultiRegionKmsMigration::UserMigrationJob to the background job schedule
A previous commit added a class for re-encrypting single-region KMS ciphertexts on user records with a multi-region key. That job was not added to the job schedule so it could be tested in lower environments before being enabled.

We have confirmed the job works as expected in lower environments.

This commit also addresses LG-11030 by making the record count and statement timeout params for the migration jobs configurable with the IdentityConfig tooling.

[skip changelog]