LG-10445: bump up the GPO verification code max attempts to 5
Most of our rate limiting allows a user 5 attempts so this brings the
entering of the GPO verification code in line with the others.

We spoke to security and confirmed that there is no issue with this
change.

changelog: User-facing Improvements, Identity Verification By Mail, Increased user's max attempts to enter their verification code before being rate limited.

Co-authored-by: Sonia Connolly <sonia.connolly@gsa.gov>