Read records in KMS migration job while timeout is in effect
Currently the ProfileMigrationJob and UserMigrationJob make expensive queries to find users to migrate. These queries need a large amount of time to run. To enable this we run them in a transaction with a local statement timeout.

I discovered an issue with this approach. The query was not actually executed in the transaction block. The transaction returned a relation and the query was executed when we attempted to iterated over the records in the relation. This commit adds a `#to_a` call to the relation to ensure the query is run and the records are loaded into memory before the transaction block is complete.

[skip changelog]