Count successful phone proofing attempts towards the rate limit
We do not currently count successful proofing attempts towards the rate limit. This was done to support a feature that prevented users from being rate limited after successfully completing a step. The logic that caused that issue was addressed in #9343.

This commit starts counting successful attempts to towards the rate limit. This protects our vendors from abuse and makes it easier for us to make this step re-entrant to support the back button.

changelog: Improvements, Rate-Limiting, The proof address rate limiter was modified to rate limit on successful address proofing attempts as well as on failed address proofing attempts