LG-11190 Don't reset the SSN rate limit on successful submission
Prior to this commit we reset the ss rate limiter on success. This was done to prevent users from being rate limited after successfully completing a step. The logic that caused that issue was addressed in #9343.

This commit starts counting successful attempts to towards the rate limit. This protects our users from fraud and makes it easier for us to make this step re-entrant to support the back button.

changelog: Improvements, Rate Limiting, The proof ssn rate limiter was modified to no longer reset on successful proofing attempts that include an ssn