LG-10849 Use hybrid flow phone number in phone form
This commit uses the phone number a user used to complete the hybrid flow if we have the user has no MFA phone number. Hopefully pre-filling the number in this way will help improve success on the phone step.

changelog: User-Facing Improvements, IdV Phone Step, The hybrid flow phone is pre-filled on the phone step if the user has no MFA phone number to pre-fill.