LG-11017: Add doc type and issuing country to event
Technical Notes:
- We had to update the `if` statement in the mock `result_response` so
that if we have `classification_info` but it's not erroneous, we don't
accidentally apply the `barcode error`.
- Add a new yml file for testing the case where there is
`classification_info` that exists and is not erroneous
- had to use hash rocket syntax to keep keys as strings instead of
symbols
- We found that merging the `doc_type_and_issuing_country` after `to_h`
did not work, as it added it to the hash returned by `to_h` rather than
the `response`. By saving the output of the line as a variable and then
passing that in to analytics, we were able to fix the issue.