Start using the new enter password url
The change in #9375 renamed the review controller to the "enter password" controller. This commit introduced a new path, but did not start using it to support the 50/50 state when that change was deployed.

This commit starts using the new path but does not remove the old ones. This is also to prevent 404s in the 50/50 state. A follow up will be needed to remove the old routes after this is deployed.

[skip changelog]