Log if a user has a pending or active profile when updating their password
If a user updates their password with an active profile we re-encrypt the active profile with their password. We do not, however, re-encrypt pending profiles.

We are considering a change that would re-encrypt pending profiles if the data was decrypted and available in the session. Before we do we want to measure the impact this change would have. Currently we do not log if a user has a pendin
g profile when updating their password so measuring impace is not possible.

This commit adds the necessary logging to make a decision about the feature described above.

[skip changelog]