LG-11403 Remove the old review routes
In #9411 we moved all traffic from the old `/verify/review` path to the new `/verify/enter_password` path. We left the old paths around so that users getting redirected by old instances would not encounter a 404 error.

Once the changes in #9411 has been deployed for a day there should not be any more users visitting `/verify/review` and we can merge this change to remove it.

[skip changelog]