Don't redirect to PhoneQuestion from HybridHandoff if referer is nil
We see that referer is nil if the user tries to visit idv_link_sent_url while still on
HybridHandoff. Not sure why, but web search indicates that the user can turn off referer
in the HTTP header, which would cause a redirect loop in #maybe_redirect_for_phone_question_ab_test.
So this adds a check for nil referer in that method.

[skip changelog]

Co-authored-by: Doug Price <douglas.price@gsa.gov>
Co-authored-by: Amir Reavis-Bey <amir.reavis-bey@gsa.gov>
Co-authored-by: Jessica Dembe: <jessica.dembe@gsa.gov>