Add a title to thje in person address template
This commit changed the app to require templates to declare a title: https://github.com/18F/identity-idp/commit/4d69e9de84fdfbd44597fa6ec60a355ae448dd08

It looks like this was able to sneak through without a title declaration in the In-Person flow on in the address show template. This commit adds a title declaration to that template.

[skip changelog]