Make Pii::Cacher#fetch_string a private method
The `Pii::Cacher#fetch_string` method gets the raw PII JSON blob and returns it. This is called by the `#fetch` in the `Pii::Cacher` and converted to `Pii::Attributes`.

Outside of the `Pii::Cacher` this method is called in one place by the IdP. This is in the completions controller to get the PII which is ultimately parsed into a hash. The same can be acheived by taking the result of `Pii::Cacher#fetch` and calling `#to_h` on it. This commit does that so the `#fetch_string` method can be made part of the private API.

Making this method part of the private API will make it easier to modify the way the cacher works to support encrypting and decrypting both active and pending profiles in the future.

[skip changelog]