Modify `Pii::Cacher#save_decrypted_pii_json` to be `Pii:Cacher#save_decrypted_pii` (#9498)
The `#save_decrypted_pii_json` method consumed a JSON PII object and wrote it into the session under the `decrypted_pii` key.

This method was called in 2 places:

- The `ReactivateAccountController`
- The `Idv::Session`

In `ReactivateAccountController` the PII is available as `Pii::Attributes` and is converted to JSON to enable a call to `#save_decrypted_pii_json`.

In the `Idv::Session` the PII is available as `Pii::Attributes`, but prior to this commit underwent quite the Rube Goldberg process to find it's way into a JSON string that was then written to the session  by `save_decrypted_pii_json`.

This commit changes `save_decrypted_pii_json` to be `save_decrypted_pii` and take a `Pii::Attributes` instead of a JSON string argument. This will make it easier to implement a version of this method that handles multiple profiles when we start encrypting both the active and pending profile in the session.

This commit also short circuits the Rube Goldberg machine in `Idv::Session` and simply writes the attributes to the user session when the profile is created.

changelog: Internal, PII session encryption, The Pii::Cacher#save_decrypted_pii_json was modified to be the Pii:Cacher#save_decrypted_pii.